Print all command (#16)
* Create 'all' command to print all log entries

Added the all command which will print all the log entries. In
order to do this the log loading functionality has been moved to it's
own function.

* Remove log data global variable

No need for log data to be a global variable within the current
application structure. The function 'loadLog' now creates and returns
the log data structure. This way the data can be loaded where needed.

* Add os.exit to the end of the all command

Added os.exit(0) to the end of the all command functionality
within the switch statement. This ensures the program will exit with
a success code after all the log messages have been printed and no
other uninteded actions can occur.

* Change beaconLogData definition to be more idiomatic

beaconLogData is now defined in the loadLog function using the
var keyword rather than a call to 'new()'.