Created messagelog package
Created a messagelog package which will hold the types and functionality
required for handling the beacon log data.

Moved existing beacon log data types and function for reading and
printing into the new package.

Added new messagelog package to the imports list for main.go and updated
the main code to use the types and functions in the new package.