Create beaconlog package (#24)
* Created messagelog package

Created a messagelog package which will hold the types and functionality
required for handling the beacon log data.

Moved existing beacon log data types and function for reading and
printing into the new package.

Added new messagelog package to the imports list for main.go and updated
the main code to use the types and functions in the new package.

* Implement Stringer interface for package 'messagelog'

Changed the 'PrintLog()' function to return a string and renamed it
as 'String()'. This now impments the Stringer Inteface for the BeaconLog
struct so we can print it with our desired formating using a standard
call to fmt.Println().

* Add tests to messagelog package

Added a tests file for the messagelog package. Basic test written
for the Stringer interface on the BeaconLog type.

* Fix unused imports

Remove unused imported packages left over from merging master back into
create_beaconlog_package branch.